Resolve issue with char * casting in format string invocation
* Use `-[NSString initWithFormat:arguments:]` instead of trying to trick the preprocessor macros of all the platforms.
* A couple drive-by warning fixes that `-Weverything` surfaced.

+++ GoogleUtilities/Logger/GULASLLogger.m
@@ -142,7 +142,7 @@ NS_ASSUME_NONNULL_BEGIN
   dispatch_async(self.dispatchQueue, ^{
 #pragma clang diagnostic push
 #pragma clang diagnostic ignored "-Wdeprecated-declarations"  // asl is deprecated
-    asl_log(self.aslClient, NULL, level, "%s", logMsg);
+    asl_log(self.aslClient, NULL, (int)level, "%s", logMsg);
 #pragma clang diagnostic pop
   });
 }
